Made skip_line use feof().
Fixed skip_aesthetics up a bit - could technically have some problems after calling skip_line with the ungetc later, and made it so we don't put a character back if we run into EOF.

In mux_parse_pipe() I added something to skip past the delimiters. Whoops, that was a big mistake.